requirements.txt
Django==1.10.1
djangorestframework
markdown
django-filter
djangorestframework-jwt
django-cors-headers
django-anymail
django-bootstrap-form
django-bootstrap3
